+++ timer.h
@@ -4,8 +4,10 @@
 #include <string.h>
 #include <sys/time.h>
 
-unsigned long long currentSeconds() {
+#include "defs.h"
+
+ull currentSeconds() {
     timespec spec;
     clock_gettime(CLOCK_THREAD_CPUTIME_ID, &spec);
-    return (unsigned long long)(static_cast<float>(spec.tv_sec) * 1e9 + static_cast<float>(spec.tv_nsec));
+    return spec.tv_sec * 1e9 + spec.tv_nsec;
 }
